+++ ScrollBooker/BusinessLogic/Booking/Employee/Domain/Model/Employee.swift
@@ -14,7 +14,7 @@ struct Employee: Identifiable, Equatable, Hashable, Sendable {
     let avatar: String?
     let job: String
     let hireDate: String
-    let ratingsAverage: Double
+    let ratingsAverage: Float
     let followersCount: Int
     let ratingsCount: Int
     let productsCount: Int
